# src/main/resources/META-INF/resources/api/openapi.yml
openapi: 3.0.3
info:
  title: Islandr External API
  description: >
    Automation-facing API for scripts, CI, and infrastructure-as-code
    (issue #15). Deliberately separate from the internal `/api/v1` surface
    the Admin Console's own UI uses — that one is free to change shape as
    the UI evolves; this one is a stability-committed contract.


    Authenticated with an API key (Admin Console → API Keys → Create),
    never a session cookie. See ADR-0026 for the full rationale, including
    why this file is hand-written/offline-generated rather than produced by
    a runtime OpenAPI extension.
  version: "1.0.0"
  license:
    name: EUPL-1.2
    url: https://joinup.ec.europa.eu/collection/eupl/eupl-text-eupl-12
servers:
  - url: /api/external/v1
    description: Path is relative to wherever your Islandr instance is reachable.
security:
  - bearerAuth: []
...
paths:
...
  /users:
    get:
      operationId: listUsers
      summary: List all users
      tags: [users]
      responses:
        "200":
          description: Every user, newest first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /users/{id}/enabled:
    put:
      operationId: setUserEnabled
      summary: Enable or disable a user
      description: >
        Disabling withdraws network access, not just the login: the user's
        peers are switched off with the account. Re-enabling does **not**
        switch them back on — some may have been off for unrelated reasons
        before the account was locked, so letting a device reconnect stays an
        explicit, per-peer action via `/peers/{id}/enabled`.
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnabledRequest"
      responses:
        "200":
          description: The user after the change. Setting the value it already has is a no-op.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: No user with that id.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        An API key created in the Admin Console (API Keys page). Sent as
        `Authorization: Bearer islandr_<...>`. Reaches only the operations
        its granted scopes cover (apikey-scopes) — a key with no special
        scope reaches this facade and nothing else; see /llms.txt for the
        scope each endpoint below requires. `full` reproduces the old,
        unscoped admin-equivalent behavior (ADR-0026, R-184).
  responses:
    Unauthorized:
      description: Missing, invalid, or revoked API key.
  schemas:
...
    EnabledRequest:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
        reason:
          type: string
          description: >
            Optional free-text context for the audit log (peers only). Never
            stored on the peer itself.
      example:
        enabled: false
        reason: offboarded
...
    User:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        nickname: { type: string, nullable: true }
        displayName: { type: string, description: "nickname if set, else name" }
        email: { type: string }
        enabled: { type: boolean }
        isAdmin: { type: boolean }
        preferredLocale: { type: string, nullable: true, enum: [de, en] }
        peerCount: { type: integer }
        # Access deadline (issue #53). Null means no expiry, which is the
        # default. Once it passes the user cannot log in and their devices are
        # switched off; accessExpired is that comparison, already made.
        validUntil: { type: string, format: date-time, nullable: true }
        accessExpired: { type: boolean }
        createdAt: { type: string, format: date-time }
...